ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
faiss-cpu>=1.7.0
sentence-transformers==3.1.1
PyPDF2==3.0.1
python-multipart==0.0.12
psycopg2-binary==2.9.9
python-dotenv==1.0.1
numpy==1.26.4
